改了一下show table😵‍💫

+++ src/main/java/cn/edu/thssdb/schema/Database.java
@@ -75,13 +75,11 @@ public class Database {
     // TODO 需要修改
     try {
       lock.writeLock().lock();
-      if (!tables.containsKey(name))
-        throw new KeyNotExistException();
+      if (!tables.containsKey(name)) throw new KeyNotExistException();
       // TODO throw new TableNotExistException(name);
       String metaFilename = storage_dir + "meta_" + this.name + "_" + name + ".data";
       File metaFile = new File(metaFilename);
-      if (metaFile.isFile())
-        metaFile.delete();
+      if (metaFile.isFile()) metaFile.delete();
       Table table = tables.get(name);
       table.dropSelf();
       tables.remove(name);
@@ -93,18 +91,10 @@ public class Database {
   // 展示database中tableName表中的元数据
   public String show(String tableName) {
     try {
-      lock.writeLock().lock();
-      final String filenamePrefix = storage_dir + "meta_" + this.name + "_";
-      final String filenameSuffix = ".data";
-      for (Table table : tables.values()) {
-        File metaFile = new File(filenamePrefix + table.tableName + filenameSuffix);
-        if (metaFile.isFile())
-          metaFile.delete();
-        table.dropSelf();
-        // tables.remove(table.tableName);
-      }
-      tables.clear();
-      tables = null;
+      lock.readLock().lock();
+      if (!tables.containsKey(tableName)) throw new KeyNotExistException();
+      Table table = tables.get(tableName);
+      return table.show();
     } finally {
       lock.readLock().unlock();
     }
